backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
anthropic==0.40.0
langgraph==0.2.59
langgraph-checkpoint-sqlite==2.0.8
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
rapidfuzz==3.10.1
aiosqlite==0.20.0
